Name every aggregate field, and verify against CI's actual warning configuration
Seven CI jobs failed on 35eddc5 with -Wmissing-field-initializers: a designated
initializer for Run omitted three of six fields. I had reported the local GCC gate
green minutes earlier.

The compiler had already told me. GCC 16 emitted that exact warning twice, into a
log on this machine. My GCC tree was configured without DW_WARNINGS_AS_ERRORS --
which defaults OFF in CMakeLists.txt and which CI sets ON -- so the diagnostic came
out as a warning rather than an error, and I grepped the log for "error:" and saw
none. The failure was not the toolchain's silence.

This is the third variant of one mistake across two phases: reading the
measurement I expected instead of the one in front of me. Phase 5 was a partial
lint log read as a complete one. This was a warning stream filtered out of
existence by the grep that was supposed to inspect it.

Fixed structurally rather than by resolve. The local GCC tree is now configured
exactly as CI configures it:

    CC=gcc-16 CXX=g++-16 cmake -S . -B /tmp/gccbuild -G Ninja \
        -DCMAKE_BUILD_TYPE=Debug -DDW_BUILD_TESTS=ON -DDW_WARNINGS_AS_ERRORS=ON

and verdicts are read from exit codes, because under warnings-as-errors the
warning/error distinction is precisely what hid this.

It paid immediately. The strict build found a second defect that Clang cannot
diagnose at all: -Wuseless-cast on static_cast<std::int64_t>(std::llround(...)),
where llround already returns long long. The cast is a no-op on this platform and
a real conversion where int64_t is long, so the destination type does the work
instead. That had been sitting in code which passed every check I had run.

Run now names all six fields at both constructions. Two toolchains disagree about
the alternative -- clang-tidy's modernize-use-designated-initializers wants the
designated form, GCC 13 warns when a designated list omits a member, and GCC 16 no
longer emits that warning -- so naming every field satisfies both and depends on
neither's version.

Local matrix, every gate by exit code:
  gcc-16 strict (CI's configuration)  542/542, exit 0
  clang release                       542/542, exit 0
  sanitize (ASan+UBSan)               542/542, exit 0
  quantlib cross-validation           5/5, exit 0
  include hygiene                     exit 0
  clang-format + clang-tidy           exit 0, 34/34 units, 0 findings, marker present
  EXP-06                              pass

+++ tests/engines/finite_difference_test.cpp
@@ -222,7 +222,7 @@ TEST(FiniteDifferenceTest, StabilityBoundIsSufficientNotNecessary) {
         const double dtau_max = (1.0 / 1000.0) / s.value().diagnostics.explicit_stability_ratio;
 
         PdeConfig config = probe;
-        config.time_steps = static_cast<std::int64_t>(std::llround(1.0 / (ratio * dtau_max)));
+        config.time_steps = std::llround(1.0 / (ratio * dtau_max));
         return FiniteDifferenceEngine::price(market, option, model, config);
     };
